+++ evalml/tests/utils_tests/test_nullable_type_utils.py
@@ -199,7 +199,6 @@ def test_downcast_nullable_y_replaces_nullable_types(
     nullable_ltype,
     has_nans,
 ):
-    # --> remove this comment
     y = nullable_type_target(ltype=nullable_ltype, has_nans=has_nans)
 
     y_d = _downcast_nullable_y(